secondary-fetch
allow to work with an additional repo for inhouse plugins
performs a first aql towards the jenkins repo to find the official
plugins and the wars.
Afterwards performs a second aql with a custom query read from a file

All file requests are done towards the second internal repository to
reduce load on the jenkins repo.

prefetch the description in parallelStream to speed up processing

add support openresty by generating a lua script instead of .htaccess
rules for apache